website/forms.py
from captcha.fields import CaptchaField
from .models import Contact, Newsletter
from django import forms


class ContactForm(forms.ModelForm):
    captcha = CaptchaField()
    class Meta:
        model = Contact
        fields = '__all__'

class NewsletterForm(forms.ModelForm):
    class Meta:
        model = Newsletter
        fields = '__all__'        