requirements.txt
asgiref==3.5.2
Django==3.2.14
pytz==2022.1
sqlparse==0.4.2
Pillow==9.2.0
django-taggit==3.0.0
django-environ==0.9.0
django-simple-captcha==0.5.17
